Create vending machine
This C program simulates a basic vending machine. It displays a menu of 7 items (like Lays chips, Jim Jam, Mazza, KitKat, etc.) along with their prices. The user is prompted to enter the item number they wish to purchase and the quantity. Based on the input, the program uses a switch statement to calculate the total cost by multiplying the item price with the quantity. If an invalid item number is entered, it shows an error message. Otherwise, it displays a message of appreciation and the total amount to be paid.